Use plone.app.testing
